# src/docs/apidocs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "E-commerce API",
    "version": "1.0.0",
    "description": "An E-commerce store with Express"
  },
  "paths": {
...
    "/api/carts": {
      "get": {
        "summary": "Get all carts",
        "operationId": "getCarts",
        "responses": {
          "200": {
            "description": "Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "timestamp": {
                        "type": "number"
                      },
                      "products": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "name": {
                              "type": "string"
                            },
                            "description": {
                              "type": "string"
                            },
                            "code": {
                              "type": "string"
                            },
                            "thumbnail": {
                              "type": "string"
                            },
                            "price": {
                              "type": "number"
                            },
                            "stock": {
                              "type": "number"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "number"
                    }
                  },
                  "default": {
                    "message": "Not authorized for the request",
                    "code": 401
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "number"
                    }
                  },
                  "default": {
                    "message": "Server encountered a problem",
                    "code": 500
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a new cart",
        "operationId": "createCart",
        "responses": {
          "200": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "number"
                    }
                  },
                  "default": {
                    "message": "Not authorized for the request",
                    "code": 401
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "number"
                    }
                  },
                  "default": {
                    "message": "Server encountered a problem",
                    "code": 500
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/carts/{id}": {
      "get": {
        "summary": "Get one cart",
        "operationId": "getCart",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "number"
                    },
                    "products": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "code": {
                            "type": "string"
                          },
                          "thumbnail": {
                            "type": "string"
                          },
                          "price": {
                            "type": "number"
                          },
                          "stock": {
                            "type": "number"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "number"
                    }
                  },
                  "default": {
                    "message": "Not authorized for the request",
                    "code": 401
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "number"
                    }
                  },
                  "default": {
                    "message": "Object not found in database",
                    "code": 404
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "number"
                    }
                  },
                  "default": {
                    "message": "Server encountered a problem",
                    "code": 500
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete one cart",
        "operationId": "deleteCart",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "number"
                    },
                    "products": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "code": {
                            "type": "string"
                          },
                          "thumbnail": {
                            "type": "string"
                          },
                          "price": {
                            "type": "number"
                          },
                          "stock": {
                            "type": "number"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "number"
                    }
                  },
                  "default": {
                    "message": "Not authorized for the request",
                    "code": 401
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "number"
                    }
                  },
                  "default": {
                    "message": "Object not found in database",
                    "code": 404
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "number"
                    }
                  },
                  "default": {
                    "message": "Server encountered a problem",
                    "code": 500
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/carts/{id}/products": {
      "get": {
        "summary": "Get products of a cart",
        "operationId": "getCartProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "All products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "products": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "code": {
                            "type": "string"
                          },
                          "thumbnail": {
                            "type": "string"
                          },
                          "price": {
                            "type": "number"
                          },
                          "stock": {
                            "type": "number"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "number"
                    }
                  },
                  "default": {
                    "message": "Not authorized for the request",
                    "code": 401
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "number"
                    }
                  },
                  "default": {
                    "message": "Object not found in database",
                    "code": 404
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "number"
                    }
                  },
                  "default": {
                    "message": "Server encountered a problem",
                    "code": 500
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BasicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    }
  },
  "security": [
    {
      "BasicAuth": []
    }
  ]
}